server/requirements.txt
dtcloud @ git+https://gitee.com/dtcloud360/dtcloud.git@DTCloud-4.19
bcrypt>=4.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
psycopg[binary]>=3.1.18
SQLAlchemy>=2.0.25
cryptography>=42.0.0
openpyxl>=3.1.2
pymupdf>=1.24.0
xlrd>=2.0.1
requests>=2.31.0
redis>=5.0.0

# 为确保 Python 3.8 可用，避免安装到后续可能移除 3.8 支持的版本
uvicorn>=0.27.0,<0.42
fastapi>=0.110.0,<0.111
python-multipart>=0.0.6
